+++ rust/main/chains/hyperlane-ethereum/src/rpc_clients/provider.rs
@@ -11,6 +11,7 @@ use ethers::{prelude::Middleware, types::TransactionReceipt};
 use ethers_contract::builders::ContractCall;
 use ethers_core::abi::Function;
 use ethers_core::types::transaction::eip2718::TypedTransaction;
+use ethers_core::types::{BlockId, FeeHistory, U256 as EthersU256};
 use ethers_core::{abi::Address, types::BlockNumber};
 use hyperlane_core::{ethers_core_types, ChainInfo, HyperlaneCustomErrorWrapper, H512, U256};
 use tokio::time::sleep;
@@ -24,6 +25,24 @@ use hyperlane_core::{
 use crate::{
     get_finalized_block_number, BuildableWithProvider, ConnectionConf, EthereumReorgPeriod,
 };
+
+// From
+// gas_limit: QUANTITY, 32 bytes - The maximum amount of gas that can be used.
+// max_fee_per_gas: QUANTITY, 32 bytes - The maximum fee per unit of gas that the sender is willing to pay.
+// max_priority_fee_per_gas: QUANTITY, 32 bytes - The maximum priority fee per unit of gas to incentivize miners.
+// gas_per_pubdata_limit: QUANTITY, 32 bytes - The gas limit per unit of public data.
+/// Response from the zkSync estimate fee endpoint.
+#[derive(Debug, Clone, PartialEq, Eq, serde::Deserialize, serde::Serialize)]
+pub struct ZksyncEstimateFeeResponse {
+    /// Gas limit
+    pub gas_limit: EthersU256,
+    /// Max fee
+    pub max_fee_per_gas: EthersU256,
+    /// Max priority fee
+    pub max_priority_fee_per_gas: EthersU256,
+    /// Gas per pubdata limit
+    pub gas_per_pubdata_limit: EthersU256,
+}
 
 /// Connection to an ethereum provider. Useful for querying information about
 /// the blockchain.
@@ -96,6 +115,26 @@ pub trait EvmProviderForSubmitter: Send + Sync {
 
     /// Read-only call into blockchain which returns a boolean
     async fn check(&self, tx: &TypedTransaction, function: &Function) -> ChainResult<bool>;
+
+    /// Get the next nonce to use for a given address (using the finalized block)
+    async fn get_next_nonce_on_finalized_block(&self, address: &Address) -> ChainResult<U256>;
+
+    /// Get the fee history
+    async fn fee_history(
+        &self,
+        block_count: U256,
+        last_block: BlockNumber,
+        reward_percentiles: &[f64],
+    ) -> ChainResult<FeeHistory>;
+
+    /// Estimate the fee for a zkSync transaction
+    async fn zk_estimate_fee(
+        &self,
+        tx: &TypedTransaction,
+    ) -> ChainResult<ZksyncEstimateFeeResponse>;
+
+    /// Get default sender
+    fn default_sender(&self) -> Option<Address>;
 }
 
 #[async_trait]
@@ -159,6 +198,41 @@ where
             .map_err(|e| ChainCommunicationError::CustomError(e.to_string()))?;
 
         Ok(success)
+    }
+
+    async fn get_next_nonce_on_finalized_block(&self, address: &Address) -> ChainResult<U256> {
+        self.provider
+            .get_transaction_count(*address, Some(BlockId::Number(BlockNumber::Finalized)))
+            .await
+            .map_err(ChainCommunicationError::from_other)
+            .map(Into::into)
+    }
+
+    async fn fee_history(
+        &self,
+        block_count: U256,
+        last_block: BlockNumber,
+        reward_percentiles: &[f64],
+    ) -> ChainResult<FeeHistory> {
+        self.provider
+            .fee_history(block_count, last_block, reward_percentiles)
+            .await
+            .map_err(ChainCommunicationError::from_other)
+    }
+
+    async fn zk_estimate_fee(
+        &self,
+        tx: &TypedTransaction,
+    ) -> ChainResult<ZksyncEstimateFeeResponse> {
+        self.provider
+            .provider()
+            .request("zks_estimateFee", [tx.clone()])
+            .await
+            .map_err(ChainCommunicationError::from_other)
+    }
+
+    fn default_sender(&self) -> Option<Address> {
+        self.provider.default_sender()
     }
 }
 
@@ -322,6 +396,13 @@ pub struct SubmitterProviderBuilder {}
 impl BuildableWithProvider for SubmitterProviderBuilder {
     type Output = Box<dyn EvmProviderForSubmitter>;
     const NEEDS_SIGNER: bool = true;
+
+    // the submitter does not use the ethers submission middleware.
+    // it uses its own logic for setting transaction parameters
+    // and landing them onchain
+    fn uses_ethers_submission_middleware(&self) -> bool {
+        false
+    }
 
     async fn build_with_provider<M: Middleware + 'static>(
         &self,