+++ typescript/cli/src/commands/warp.ts
@@ -4,18 +4,18 @@ import { CommandModule } from 'yargs';
 import {
   ChainName,
   ChainSubmissionStrategySchema,
+  expandVirtualWarpDeployConfig,
   expandWarpDeployConfig,
   getRouterAddressesFromWarpCoreConfig,
 } from '@hyperlane-xyz/sdk';
 import { ProtocolType, assert, objFilter } from '@hyperlane-xyz/utils';
 
 import { runWarpRouteCheck } from '../check/warp.js';
-import {
-  createWarpRouteDeployConfig,
-  readWarpRouteDeployConfig,
-} from '../config/warp.js';
+import { createWarpRouteDeployConfig } from '../config/warp.js';
 import {
   CommandModuleWithContext,
+  CommandModuleWithWarpApplyContext,
+  CommandModuleWithWarpDeployContext,
   CommandModuleWithWriteContext,
 } from '../context/types.js';
 import { evaluateIfDryRunFailure } from '../deploy/dry-run.js';
@@ -75,10 +75,11 @@ export const warpCommand: CommandModule = {
   handler: () => log('Command required'),
 };
 
-export const apply: CommandModuleWithWriteContext<{
-  config: string;
+export const apply: CommandModuleWithWarpApplyContext<{
+  config?: string;
+  warp?: string;
   symbol?: string;
-  warp: string;
+  warpRouteId?: string;
   strategy?: string;
   receiptsDir: string;
 }> = {
@@ -86,12 +87,13 @@ export const apply: CommandModuleWithWriteContext<{
   describe: 'Update Warp Route contracts',
   builder: {
     config: warpDeploymentConfigCommandOption,
-    symbol: {
-      ...symbolCommandOption,
-      demandOption: false,
-    },
+    warpRouteId: warpRouteIdCommandOption,
     warp: {
       ...warpCoreConfigCommandOption,
+      demandOption: false,
+    },
+    symbol: {
+      ...symbolCommandOption,
       demandOption: false,
     },
     strategy: { ...strategyCommandOption, demandOption: false },
@@ -102,33 +104,17 @@ export const apply: CommandModuleWithWriteContext<{
       coerce: (dir) => removeEndingSlash(dir),
     },
   },
-  handler: async ({
-    context,
-    config,
-    symbol,
-    warp,
-    strategy: strategyUrl,
-    receiptsDir,
-  }) => {
+  handler: async ({ context, strategy: strategyUrl, receiptsDir }) => {
     logCommandHeader('Hyperlane Warp Apply');
-
-    const warpCoreConfig = await getWarpCoreConfigOrExit({
-      symbol,
-      warp,
-      context,
-    });
 
     if (strategyUrl)
       ChainSubmissionStrategySchema.parse(readYamlOrJson(strategyUrl));
-    const warpDeployConfig = await readWarpRouteDeployConfig({
-      filePath: config,
-      context,
-    });
 
     await runWarpRouteApply({
       context,
-      warpDeployConfig,
-      warpCoreConfig,
+      // Already fetched in the resolveWarpApplyChains
+      warpDeployConfig: context.warpDeployConfig,
+      warpCoreConfig: context.warpCoreConfig,
       strategyUrl,
       receiptsDir,
     });
@@ -136,10 +122,12 @@ export const apply: CommandModuleWithWriteContext<{
   },
 };
 
-export const deploy: CommandModuleWithWriteContext<{
-  config: string;
+export const deploy: CommandModuleWithWarpDeployContext<{
+  config?: string;
   'dry-run': string;
   'from-address': string;
+  symbol?: string;
+  warpRouteId?: string;
 }> = {
   command: 'deploy',
   describe: 'Deploy Warp Route contracts',
@@ -147,8 +135,13 @@ export const deploy: CommandModuleWithWriteContext<{
     config: warpDeploymentConfigCommandOption,
     'dry-run': dryRunCommandOption,
     'from-address': fromAddressCommandOption,
+    symbol: {
+      ...symbolCommandOption,
+      demandOption: false,
+    },
+    warpRouteId: warpRouteIdCommandOption,
   },
-  handler: async ({ context, config, dryRun }) => {
+  handler: async ({ context, dryRun }) => {
     logCommandHeader(
       `Hyperlane Warp Route Deployment${dryRun ? ' Dry-Run' : ''}`,
     );
@@ -156,7 +149,8 @@ export const deploy: CommandModuleWithWriteContext<{
     try {
       await runWarpRouteDeploy({
         context,
-        warpRouteDeploymentConfigPath: config,
+        // Already fetched in the resolveWarpRouteConfigChains
+        warpDeployConfig: context.warpDeployConfig,
       });
     } catch (error: any) {
       evaluateIfDryRunFailure(error, dryRun);
@@ -178,7 +172,7 @@ export const init: CommandModuleWithContext<{
       describe: 'Create an advanced ISM',
       default: false,
     },
-    out: outputFileCommandOption(DEFAULT_WARP_ROUTE_DEPLOYMENT_CONFIG_PATH),
+    out: outputFileCommandOption(),
   },
   handler: async ({ context, advanced, out }) => {
     logCommandHeader('Hyperlane Warp Configure');
@@ -389,13 +383,8 @@ export const check: CommandModuleWithContext<{
       warpCoreConfig,
     ));
 
-    // Expand the config before removing non-EVM chain configs to correctly expand
-    // the remote routers
-    let expandedWarpDeployConfig = await expandWarpDeployConfig(
-      context.multiProvider,
-      warpDeployConfig,
-      getRouterAddressesFromWarpCoreConfig(warpCoreConfig),
-    );
+    const deployedRoutersAddresses =
+      getRouterAddressesFromWarpCoreConfig(warpCoreConfig);
 
     // Remove any non EVM chain configs to avoid the checker crashing
     warpCoreConfig.tokens = warpCoreConfig.tokens.filter(
@@ -404,20 +393,33 @@ export const check: CommandModuleWithContext<{
         ProtocolType.Ethereum,
     );
 
-    expandedWarpDeployConfig = objFilter(
-      expandedWarpDeployConfig,
-      (chain, _config): _config is any =>
-        context.multiProvider.getProtocol(chain) === ProtocolType.Ethereum,
-    );
-
     // Get on-chain config
     const onChainWarpConfig = await getWarpRouteConfigsByCore({
       context,
       warpCoreConfig,
     });
 
-    await runWarpRouteCheck({
+    // get virtual on-chain config
+    const expandedOnChainWarpConfig = await expandVirtualWarpDeployConfig({
+      multiProvider: context.multiProvider,
       onChainWarpConfig,
+      deployedRoutersAddresses,
+    });
+
+    let expandedWarpDeployConfig = await expandWarpDeployConfig({
+      multiProvider: context.multiProvider,
+      warpDeployConfig,
+      deployedRoutersAddresses,
+      expandedOnChainWarpConfig,
+    });
+    expandedWarpDeployConfig = objFilter(
+      expandedWarpDeployConfig,
+      (chain, _config): _config is any =>
+        context.multiProvider.getProtocol(chain) === ProtocolType.Ethereum,
+    );
+
+    await runWarpRouteCheck({
+      onChainWarpConfig: expandedOnChainWarpConfig,
       warpRouteConfig: expandedWarpDeployConfig,
     });